Move the test_agent_otlp_upload from the test/otel_tracing_e2e/test_e2e.py::Test_OTelMetricE2E class to a new tests/otel/test_metrics.py::Test_OTelMetrics class
